refactor(support): split tunnel-client into packages and add configurable rate limits
Refactor the tunnel-client from a single 1181-line main.go into organized
internal packages (config, connection, discovery, models, stream, terminal).
Rename traefik.go to nethserver.go with updated function names and log messages.
Replace YAML config with EXCLUDE_PATTERNS env var / --exclude flag for service
filtering. Improve api-cli error logging to include stderr output. Add
configurable rate limiting via env vars (RATE_LIMIT_TUNNEL_PER_IP,
RATE_LIMIT_TUNNEL_PER_KEY, RATE_LIMIT_SESSION_PER_ID, RATE_LIMIT_WINDOW)
with session limit raised from 100 to 500 req/min. Add build-tunnel-client
and run-tunnel-client Makefile targets.
