+++ services/support/.env.example
@@ -25,6 +25,12 @@ TUNNEL_GRACE_PERIOD=2m
 MAX_TUNNELS=1000
 MAX_SESSIONS_PER_SYSTEM=5
 
+# Rate limiting
+# RATE_LIMIT_TUNNEL_PER_IP=10
+# RATE_LIMIT_TUNNEL_PER_KEY=5
+# RATE_LIMIT_SESSION_PER_ID=500
+# RATE_LIMIT_WINDOW=1m
+
 # Internal authentication (shared secret with backend)
 # IMPORTANT: Generate with: openssl rand -hex 32
 INTERNAL_SECRET=CHANGE_ME_GENERATE_WITH_openssl_rand_hex_32
